move getAllValidatorsFeed to common
