+++ api/lib/reducers/common.js
@@ -27,3 +27,30 @@ module.exports.getRanksForValidators = function getRanksForValidators(
       rank: ++index
     }))
 }
+
+module.exports.getAllValidatorsFeed = async function getAllValidatorsFeed(
+  validators,
+  allValidatorsAddresses,
+  networkList,
+  dataSource,
+  network
+) {
+  const allValidatorsFeed = await dataSource.db.getAccountsNotifications(
+    allValidatorsAddresses,
+    network.id
+  )
+  return validators.map((validator) => {
+    const validatorFeed = allValidatorsFeed.filter(
+      ({ resourceId }) => resourceId === validator.operatorAddress
+    )
+    return {
+      ...validator,
+      feed:
+        validatorFeed && Array.isArray(validatorFeed)
+          ? validatorFeed.map((notification) =>
+              dataSource.reducers.notificationReducer(notification, networkList)
+            )
+          : []
+    }
+  })
+}
